Tmin is now stored for each data type seperately
Therefore we can e.g. set
tmin for residuals to 0
and for forces to 1.23434

statistics_text_field is now an derived class which allows
for handling a dataframe. So we have full control on the
output and can also make use of pandas built-in export function.
When exporting/saving the statistics data, user can decide
between plain text (as shown in the output window) or csv-format